# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NiazFinder API",
    "version": "0.2.x-unstable",
    "description": "Generated skeleton of the Next.js App Router API surface. Methods and auth requirements are statically detected; only operations with explicit request/response schemas are stable. Regenerate with `npm run docs:openapi`.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/sasanappstore2/NiazFinder-main/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev (`npm run dev`)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "User token from phone-OTP login (AuthToken)."
      },
      "internalSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret",
        "description": "INTERNAL_API_SECRET for worker/cron routes. Constant-time compare, fail-closed."
      }
    }
  },
  "paths": {
...
    "/api/super-admin/categories/{id}": {
      "patch": {
        "tags": [
          "super-admin"
        ],
        "summary": "PATCH /api/super-admin/categories/[id]",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      },
      "delete": {
        "tags": [
          "super-admin"
        ],
        "summary": "DELETE /api/super-admin/categories/[id]",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      }
    },
...
    "/api/super-admin/filing-scrapers": {
      "get": {
        "tags": [
          "super-admin"
        ],
        "summary": "GET /api/super-admin/filing-scrapers",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      },
      "post": {
        "tags": [
          "super-admin"
        ],
        "summary": "POST /api/super-admin/filing-scrapers",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      }
    },
    "/api/super-admin/filing-scrapers/{id}": {
      "patch": {
        "tags": [
          "super-admin"
        ],
        "summary": "PATCH /api/super-admin/filing-scrapers/[id]",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      },
      "delete": {
        "tags": [
          "super-admin"
        ],
        "summary": "DELETE /api/super-admin/filing-scrapers/[id]",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      }
    },
...
    "/api/telemetry/post-intake": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "POST /api/telemetry/post-intake",
        "description": "No server-side auth call detected statically — verify before treating as public.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      }
    },
...
  }
}